week9_final_project/requirements.txt
sys
os
requests
freecurrencyapi
python-dotenv
countryinfo
iso639-lang
iso3166